+++ lib/lightning_web/live/auth_providers/form_component.ex
@@ -267,7 +267,7 @@ defmodule LightningWeb.AuthProvidersLive.FormComponent do
                     external authentication for Lightning.
                   </p>
 
-                  <Common.button color="red" phx-click="disable" phx_target={@myself}>
+                  <Common.button color="red" phx-click="disable" phx-target={@myself}>
                     <div class="h-full">
                       <Heroicons.trash class="h-4 w-4 inline-block" />
                       <span class="inline-block align-middle">Remove</span>